Refactor code and add commentscounter test
