+++ package.json
@@ -4,9 +4,23 @@
   "description": "",
   "private": true,
   "scripts": {
-    "test": "jest --coverage --watch",
+    "test": "jest --watch",
     "build": "webpack",
     "start": "webpack serve --open"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "jsx"
+    ],
+    "moduleDirectories": [
+      "node_modules",
+      "bower_components",
+      "src"
+    ],
+    "moduleNameMapper": {
+      ".+\\.(css|styl|less|sass|scss|png|jpg|ttf|svg|woff2)$": "identity-obj-proxy"
+    }
   },
   "repository": {
     "type": "git",
@@ -28,7 +42,10 @@
     "eslint-plugin-import": "^2.25.3",
     "hint": "^6.1.9",
     "html-webpack-plugin": "^5.5.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^27.3.1",
+    "jsdom": "^18.1.0",
+    "node-fetch": "^3.1.0",
     "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
@@ -39,6 +56,7 @@
     "webpack-dev-server": "^4.3.1"
   },
   "dependencies": {
+    "cross-fetch": "^3.1.4",
     "lodash": "^4.17.21"
   }
 }
